requirements.txt
Click==7.0
ImageHash==4.0
numpy==1.16.4
Pillow==6.1.0
PyWavelets==1.0.3
scipy==1.2.2
six==1.4.1
